Refactor TUI session modal into modular components
